Separate instance related variables into its own base class
